+++ include/common/monolish_common.hpp
@@ -32,7 +32,8 @@ namespace util {
  *    - # of data transfer: 0
  */
 template <typename T>
-T get_residual_l2(matrix::CRS<T> &A, vector<T> &x, vector<T> &b);
+T get_residual_l2(const matrix::CRS<T> &A, const vector<T> &x,
+                  const vector<T> &b);
 
 /**
  * @brief get nrm |b-Ax|_2
@@ -46,7 +47,8 @@ T get_residual_l2(matrix::CRS<T> &A, vector<T> &x, vector<T> &b);
  *    - # of data transfer: 0
  */
 template <typename T>
-T get_residual_l2(matrix::Dense<T> &A, vector<T> &x, vector<T> &b);
+T get_residual_l2(const matrix::Dense<T> &A, const vector<T> &x,
+                  const vector<T> &b);
 
 /**
  * @brief get nrm |b-Ax|_2
@@ -59,7 +61,8 @@ T get_residual_l2(matrix::Dense<T> &A, vector<T> &x, vector<T> &b);
  * - GPU acceleration: false
  */
 template <typename T>
-T get_residual_l2(matrix::LinearOperator<T> &A, vector<T> &x, vector<T> &b);
+T get_residual_l2(const matrix::LinearOperator<T> &A, const vector<T> &x,
+                  const vector<T> &b);
 
 /**
  * @brief check error
@@ -85,13 +88,13 @@ bool solver_check(const int err);
  * 3. logging solvers and BLAS functions and utils (send, recv,
  *allocation...etc.)
  **/
-void set_log_level(size_t Level);
+void set_log_level(const size_t Level);
 
 /**
  * @brief Specifying the log finename
  * @param filename the log filename
  **/
-void set_log_filename(std::string filename);
+void set_log_filename(const std::string filename);
 
 // create typical data///////////////////////////