+++ src/utils/equation_utils.cpp
@@ -29,7 +29,8 @@ bool util::solver_check(const int err) {
 }
 
 template <typename T>
-T util::get_residual_l2(matrix::CRS<T> &A, vector<T> &x, vector<T> &b) {
+T util::get_residual_l2(const matrix::CRS<T> &A, const vector<T> &x,
+                        const vector<T> &b) {
   Logger &logger = Logger::get_instance();
   logger.util_in(monolish_func);
   vector<T> tmp(x.size());
@@ -40,12 +41,15 @@ T util::get_residual_l2(matrix::CRS<T> &A, vector<T> &x, vector<T> &b) {
   logger.util_out();
   return blas::nrm2(tmp);
 }
-template double util::get_residual_l2(matrix::CRS<double> &A, vector<double> &x,
-                                      vector<double> &b);
-template float util::get_residual_l2(matrix::CRS<float> &A, vector<float> &x,
-                                     vector<float> &b);
+template double util::get_residual_l2(const matrix::CRS<double> &A,
+                                      const vector<double> &x,
+                                      const vector<double> &b);
+template float util::get_residual_l2(const matrix::CRS<float> &A,
+                                     const vector<float> &x,
+                                     const vector<float> &b);
 template <typename T>
-T util::get_residual_l2(matrix::Dense<T> &A, vector<T> &x, vector<T> &b) {
+T util::get_residual_l2(const matrix::Dense<T> &A, const vector<T> &x,
+                        const vector<T> &b) {
   Logger &logger = Logger::get_instance();
   logger.util_in(monolish_func);
   vector<T> tmp(x.size());
@@ -56,14 +60,16 @@ T util::get_residual_l2(matrix::Dense<T> &A, vector<T> &x, vector<T> &b) {
   logger.util_out();
   return blas::nrm2(tmp);
 }
-template double util::get_residual_l2(matrix::Dense<double> &A,
-                                      vector<double> &x, vector<double> &b);
-template float util::get_residual_l2(matrix::Dense<float> &A, vector<float> &x,
-                                     vector<float> &b);
+template double util::get_residual_l2(const matrix::Dense<double> &A,
+                                      const vector<double> &x,
+                                      const vector<double> &b);
+template float util::get_residual_l2(const matrix::Dense<float> &A,
+                                     const vector<float> &x,
+                                     const vector<float> &b);
 
 template <typename T>
-T util::get_residual_l2(matrix::LinearOperator<T> &A, vector<T> &x,
-                        vector<T> &b) {
+T util::get_residual_l2(const matrix::LinearOperator<T> &A, const vector<T> &x,
+                        const vector<T> &b) {
   Logger &logger = Logger::get_instance();
   logger.util_in(monolish_func);
   vector<T> tmp(x.size());
@@ -74,9 +80,11 @@ T util::get_residual_l2(matrix::LinearOperator<T> &A, vector<T> &x,
   logger.util_out();
   return blas::nrm2(tmp);
 }
-template double util::get_residual_l2(matrix::LinearOperator<double> &A,
-                                      vector<double> &x, vector<double> &b);
-template float util::get_residual_l2(matrix::LinearOperator<float> &A,
-                                     vector<float> &x, vector<float> &b);
+template double util::get_residual_l2(const matrix::LinearOperator<double> &A,
+                                      const vector<double> &x,
+                                      const vector<double> &b);
+template float util::get_residual_l2(const matrix::LinearOperator<float> &A,
+                                     const vector<float> &x,
+                                     const vector<float> &b);
 
 } // namespace monolish